lib: remove redundant cast in acpi_verify_checksum
The (int) cast on a boolean expression assigned to an unsigned
return value serves no purpose; simplify to a direct comparison.

+++ lib/acpi.c
@@ -101,7 +101,7 @@ acpi_verify_checksum(const uint8_t *mem, size_t size)
         while (size--)
                 sum += *mem++;
 
-        return (int)(sum == 0);
+        return sum == 0;
 }
 
 /**
